Update log messages to use inttypes
Bug: 146020579
Test: ran "atest --host chre_chpp_linux_tests -c"

+++ chpp/include/chpp/app.h
@@ -163,7 +163,7 @@ struct ChppService {
   ChppDispatchFunction *dispatchFunctionPtr;
 
   //! Minimum valid length of datagrams for the service.
-  uint8_t minLength;
+  size_t minLength;
 };
 
 struct ChppAppState {
